+++ src/features/messages/components/layout/index.ts
@@ -2,5 +2,4 @@
 // Layout components barrel export
 
 export { MessagesLayout } from "./MessagesLayout";
-export { MessagesSidebar } from "./MessagesSidebar";
 export { MessagesHeader } from "./MessagesHeader";